企业折扣返回类型修改

+++ src/main/java/com/zhuanche/serv/busManage/BusAssignmentService.java
@@ -230,7 +230,7 @@ public class BusAssignmentService {
                                 }).findFirst().ifPresent(business -> {
                                     JSONObject jsonObject = (JSONObject) JSON.toJSON(business);
                                     order.setCompanyType(jsonObject.getInteger("type"));
-                                    order.setPercent(jsonObject.getInteger("percent"));
+                                    order.setPercent(jsonObject.getDouble("percent"));
                                 });
                             }
                         });
